More warning suppression.
2001-05-07  Jody Goldberg <jgoldberg@home.com>

	* ms-escher.c (ms_escher_get_data) : suppress warnings.
	(ms_escher_read_ClientTextbox) : no side effects in precondition.

	* ms-excel-read.c (ms_excel_formula_shared) : ditto.

2001-05-07  Jody Goldberg <jgoldberg@home.com>

	* src/item-cursor.c (item_cursor_event) : ignore events while editing.

	* src/functions/fn-string.c (gnumeric_value) : s/free/g_free/
	* src/parse-util.c (parse_cell_name_list) : ditto.

	* src/parser.y (gnumeric_expr_parser) : reorganize slightly to improve
	  readability.

	* src/sheet.c (sheet_range_contains_region) : doh!.  forgot to invert
	  logic when I renamed the function.

+++ plugins/excel/ms-container.c
@@ -73,7 +73,7 @@ ms_container_get_blip (MSContainer *container, int blip_id)
 		    return ms_container_get_blip (container->parent_container,
 						  blip_id);
 
-	g_return_val_if_fail (blip_id < container->blips->len, NULL);
+	g_return_val_if_fail (blip_id < (int)container->blips->len, NULL);
 
 	return g_ptr_array_index (container->blips, blip_id);
 }
